refactor(#4226): Move probe packs test to MjProbeTest and make xmir final
