+++ eo-maven-plugin/src/test/java/org/eolang/maven/MjProbeTest.java
@@ -9,20 +9,48 @@ import com.yegor256.MktmpResolver;
 import com.yegor256.WeAreOnline;
 import java.io.IOException;
 import java.nio.file.Path;
+import org.cactoos.io.InputOf;
 import org.cactoos.io.ResourceOf;
+import org.eolang.jucs.ClasspathSource;
+import org.eolang.parser.EoSyntax;
+import org.eolang.xax.XtSticky;
+import org.eolang.xax.XtYaml;
+import org.eolang.xax.XtoryMatcher;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.params.ParameterizedTest;
 
 /**
  * Test case for {@link MjProbe}.
  *
  * @since 0.28.11
+ * @todo #4226:30min Move {@link  MjProbeTest#checksProbePacks(String)} to eo-runtime.
+ *  This test checks the functionality related to the `eo-runtime` module only.
+ *  There is no need to keep it in the `eo-maven-plugin` module since it doesn't touch
+ *  the eo-maven-plugin functionality.
  */
 @ExtendWith(WeAreOnline.class)
 @ExtendWith(MktmpResolver.class)
 final class MjProbeTest {
+
+    @ParameterizedTest
+    @ClasspathSource(value = "org/eolang/maven/probe-packs/", glob = "**.yaml")
+    void checksProbePacks(final String yaml) {
+        MatcherAssert.assertThat(
+            "passed without exceptions",
+            new XtSticky(
+                new XtYaml(
+                    yaml,
+                    eo -> new EoSyntax(
+                        new InputOf(String.format("%s\n", eo))
+                    ).parsed()
+                )
+            ),
+            new XtoryMatcher()
+        );
+    }
 
     @Test
     void findsProbesViaOfflineHashFile(@Mktmp final Path temp) throws IOException {
